sticker-api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0
nltk>=3.8.1
text2emotion==0.0.5
